requirements.txt
# FastAPI & Web Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# Database
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1

# HTTP Client
httpx==0.27.2

# Google Cloud
google-cloud-storage==2.18.2
google-cloud-texttospeech==2.17.2
google-cloud-speech==2.27.0
google-auth==2.35.0
google-api-python-client==2.149.0

# Firebase
firebase-admin==6.5.0

# Validation
pydantic==2.9.2
pydantic-settings==2.6.0
email-validator==2.2.0

# Utilities
python-dateutil==2.9.0
pytz==2024.2
